// numa-admin-cli/src/types.ts
export interface CognitoConfig {
  userPoolId: string;
  clientId: string;
  region: string;
}

/**
 * Configuration for a single environment.
 */
export interface EnvironmentConfig {
  clientName: string;
  clientAccountId: string;
  region: string;
  awsProfile?: string;
  cognito?: CognitoConfig;
}